# config/openapi.yaml
...
openapi: 3.0.0
info:
  title: Portal API
  description: Portal REST API.
  version: 0.1.0
servers:
  - url: 'http://tnp-iiitr.herokuapp.com'
security:
  - bearerAuth: []
...
paths:
...
  /users/login:
    post:
      tags:
        - Users
      summary: Login
      description: Login with username and password.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  username:
                    type: string
                    minLength: 1
                  type:
                    type: string
                    minLength: 1
                  avatar:
                    type: string
                    minLength: 1
                  email:
                    type: string
                    minLength: 1
                  emailVerified:
                    type: boolean
                  expiry:
                    type: number
                  token:
                    type: string
                    minLength: 1
                required:
                  - username
                  - type
                  - avatar
                  - email
                  - emailVerified
                  - expiry
                  - token
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: Unauthorized
      requestBody:
        content:
          application/json:
            schema:
              description: ''
              type: object
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
                g_recaptcha_response:
                  type: string
                  minLength: 1
              required:
                - username
                - password
                # - g_recaptcha
        description: ''
  /users/logout:
    get:
      tags:
        - Users
      summary: Logout
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: string
              example: null
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
              example: Bad Request
  /users/signup:
    post:
      tags:
        - Users
      summary: Company SignUp
      description: Sign Up for Companies.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: jatinsachdeva@example.com
                username: jatinsachdeva
                password: '12345678'
                type: COMPANY
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                email: cs19b1003@iiitr.ac.in
                username: jatin
                type: COMPANY
                avatar: 'https://avatars0.githubusercontent.com/u/186909'
                expiry: '162683694148510800'
                token: 76ea004d-1284-41f6-9f65-a77bd1851398
                id: cd56e277-aa75-4977-9c45-87ffe7837b11
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              example:
                msg: email must be unique
            text/plain:
              schema:
                type: string
              example: Invalid body data
  /users/signup/studentwithgoogle:
    post:
      tags:
        - Users
      summary: SignUp Student
      description: Sign Up for students using Google Oauth token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                access_token: '{{access_token}}'
                password: rty890rt
                department: CSE
                year: 2019
                phone: '8888888888'
                address: |-
                  Address line.
                  Can be multiline also
                cgpa: 8.86
                semTillCGPA: 4
                resume: 'https://crossphoton.tech/cdn/CV.pdf'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                username: CS19B1003
                email: cs19b1003@iiitr.ac.in
                emailVerified: true
                type: STUDENT
                avatar: >-
                  https://lh3.googleusercontent.com/a-/AOh14GjD-BWrvQVWZKQ5N0qjam5FtuDwvZdqjo076a8TF_0=s96-c
                metaId: f33591da-1481-458b-9bc5-b91ade7eeff5
                expiry: '162683786754010800'
                token: 43d9b9c0-73e0-4e7f-ae12-a2cbca8f1548
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
              example: Invalid body data
            application/json:
              schema:
                type: object
              example:
                msg: Validation isIn on department failed
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: Invalid access_token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...